Separated library from TaskServer handlers
